Fixed pydoc/help() documentation
Reorganizing the modules into a package resulted in impoverished documentation, where functions but not classes defined in axographio.pyx were listed. These changes fix this so that help(axographio) contains essentially the same content it previously had.

The manifest was also corrected to include the moved axograph_readwrite files, as well as the license.