Made tests accessible from within the package
Before this change, test_axographio.py needed to be run on the command line as a script. These changes allow the test suite to be run using `axographio.tests.run()` inside an interactive session or other scripts.

Discoved that 50632dd, which introduced `__AXOGRAPHIO_SETUP__`, broke `python setup.py test`. This remains to be fixed.