Fixed python setup.py test
`python setup.py test` was broken by 50632dd. In that commit. a means was introduced for avoiding loading the not-yet-built extension during setup so that axographio/version.py could be loaded for fetching the git rev number. This method used a global variable, __AXOGRAPHIO_SETUP__ and had the effect of avoiding loading the extension, even if it was already built, anytime setup.py was run. This broke `python setup.py test` because the tests rely on the extension.

Here a much simpler means of fetching the git rev number from axographio/version.py is used (parse the file with exec) which avoids loading the package altogether. This obviates the need for the kludgy __AXOGRAPGIO_SETUP__ and fixes `python setup.py test`.