Removed an example notebook
The open file dialog box doesn't work on MyBinder.org. The user's data could be uploaded to the remote Jupyter server using the "Upload" button in the file browser view and then imported by specifying the path manually, but this is more complicated a procedure than I want to include in an example file.